修改WorkerId工程名

+++ duuid/workid-provider/src/main/java/com/silong/fundation/duuid/spi/WorkerIdAllocator.java
@@ -14,5 +14,5 @@ public interface WorkerIdAllocator {
    * @param info worker信息
    * @return workerid
    */
-  int allocate(WorkerInfo info);
+  long allocate(WorkerInfo info);
 }
